+++ src/rv/world/rendering/PhongWorldRenderer.java
@@ -115,6 +115,7 @@ public class PhongWorldRenderer implements SceneRenderer
 		gl.glEnable(GL.GL_BLEND);
 		if (world.getSelectedObject() != null)
 			world.getSelectedObject().renderSelected(gl);
+		world.renderBallCircle(gl);
 		if (drawings.isVisible())
 			drawings.render(gl, Renderer.glut);
 		shader.enable(gl);